Improve pattern accuracy and Interval operations.
Improved the pattern accuracy by using lookahead groups in RegEx to
certify the interval operators (SUM|DIV|SUB|MUL) are used
only with intervals (A1:5 or A1:5:2), instead of just a single cell.
And them, implemented the get_interval method, to parse the str
and Yields the values. And to map the operator with their terms, the
map_operator method was implemented.