Compute method and minor bugs with literal nums
Implemented the Compute method with support for cyclic recursion
detection with an explanatory message showing the call stack.
The minor bugs was that: the hole pattern used to check the literal
numbers (expressed directly on the formula) is accepted... But inside
the parsing process I forget to impletent the compatbility for that...
But now I implement it.