Refactor and minor fix in compute and get_value
Refactoring the compute to loop over the terms instead of repeating
code for each term (duplicating it). Now, one change in that loop
affects both terms. Another thing was an filter to prevent getting erros
if an element isnt defined on sheet, returning 0 by default. Beyond that
, if the value stored in a particular cell isnt an number, them will
raises an error.
Fixed the error message to display the call stack of cells and tell
the user how cyclic recursion occured.
Lastly, implemented the new method get_value, used inside the compute
method to get the value from another CellFormula. This allow the program
compute the value only if the value isnt setted yet (not computed).
Otherwise, it will return the stored value.