Remove CellFormula from main and minor bug fix
Finnaly the CellFormula has its own file with all their implementation.
And the minor bug fix was on the result of the operation inside the
compute method I convert the terms to integer, removing the precision
of floating points. I discover that when the DIV operator was tested
raises an ZeroDivisionError.