parse_tokens fix and finish CyclicError&CallStack
Fix a little bug found in cellformula.py at line 207, where should be
1 (size of the tokens) and there was a zero.

And I accidentally forget finish the implementation of the CallStack and
CyclicError detection. Now, using the CellCallStack namedtuple I can
keep track of the name of each cell inside the CallSack (now, a list of
CellCallStack) by setting the CellFormula instance and the label (eg A1)
and at the end display how the cyclic error occoured.